+++ miden-vm/src/cli/run.rs
@@ -6,7 +6,7 @@ use miden_core_lib::CoreLibrary;
 use miden_processor::{
     DefaultHost, ExecutionOptions, ExecutionTrace, fast::FastProcessor, parallel::build_trace,
 };
-use miden_vm::{DEFAULT_CORE_TRACE_FRAGMENT_SIZE, internal::InputFile};
+use miden_vm::internal::InputFile;
 use tracing::instrument;
 
 use super::{
@@ -154,7 +154,7 @@ fn run_masp_program(params: &RunCmd) -> Result<(ExecutionTrace, [u8; 32]), Repor
     };
 
     let (execution_output, trace_generation_context) = processor
-        .execute_for_trace_sync(&program, &mut host, DEFAULT_CORE_TRACE_FRAGMENT_SIZE)
+        .execute_for_trace_sync(&program, &mut host)
         .wrap_err("Failed to execute program")?;
 
     let trace = build_trace(
@@ -225,7 +225,7 @@ fn run_masm_program(params: &RunCmd) -> Result<(ExecutionTrace, [u8; 32]), Repor
     };
 
     let (execution_output, trace_generation_context) = processor
-        .execute_for_trace_sync(&program, &mut host, DEFAULT_CORE_TRACE_FRAGMENT_SIZE)
+        .execute_for_trace_sync(&program, &mut host)
         .wrap_err("Failed to execute program")?;
 
     let trace = build_trace(